requirements.txt
# GPU-optimized versions for CUDA support
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124

# Core ML libraries
transformers>=4.21.0
timm>=0.6.0

# Image processing and utilities
Pillow>=8.0.0
numpy>=1.21.0
requests>=2.25.0

# Optional: GPU-accelerated image processing
# opencv-python>=4.5.0  # Uncomment for advanced image processing
# cupy-cuda12x>=12.0.0  # Uncomment for GPU-accelerated NumPy operations 